+++ src/common.ts
@@ -37,16 +37,28 @@ export interface UploadDisplayDragEvents {
   onDragover?: (event: DragEvent) => void;
   onDragleave?: (event: DragEvent) => void;
 }
-/** 通用全局类型 */
+
+export type ImageEvent = Event;
+
+/**
+ * 通用全局类型
+ * */
+export type PlainObject = { [key: string]: any };
 
 export type OptionData = {
   label?: string;
   value?: string | number;
-} & { [key: string]: any };
+} & PlainObject;
 
-export type TreeOptionData = {
-  children?: Array<TreeOptionData>;
-} & OptionData;
+export type TreeOptionData<T = string | number> = {
+  children?: Array<TreeOptionData<T>>;
+  /** option label content */
+  label?: string | TNode;
+  /** option search text */
+  text?: string;
+  /** option value */
+  value?: T;
+} & PlainObject;
 
 export type SizeEnum = 'small' | 'medium' | 'large';