+++ src/progress/props.ts
@@ -12,7 +12,7 @@ export default {
   /** 进度条颜色。示例：'#ED7B2F' 或 'orange' 或 `['#f00', '#0ff', '#f0f']` 或 `{ '0%': '#f00', '100%': '#0ff' }` 或  `{ from: '#000', to: '#000' }` 等 */
   color: {
     type: [String, Object, Array] as PropType<TdProgressProps['color']>,
-    default: '',
+    default: '' as TdProgressProps['color'],
   },
   /** 进度百分比，可自定义 */
   label: {